backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.7.4
pydantic-settings==2.1.0
python-dotenv==1.0.0
openai==1.12.0
google-genai
groq==0.4.2

langchain==0.1.9
langchain-community==0.0.21
langchain-core==0.1.28
langchain-text-splitters==0.0.1
langchain-openai==0.0.5

faiss-cpu==1.8.0
sentence-transformers==2.3.1

python-multipart==0.0.9
aiofiles==23.2.1
